+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/mcpdirectstudioapp/App.kt
@@ -1,16 +1,13 @@
 package ai.mcpdirect.mcpdirectstudioapp
 
+//import ai.mcpdirect.studio.app.tool.ToolDetailScreen
 import ai.mcpdirect.studio.app.NavigationTopBar
 import ai.mcpdirect.studio.app.Screen
 import ai.mcpdirect.studio.app.UIState
 import ai.mcpdirect.studio.app.agent.MyStudioScreen
-import ai.mcpdirect.studio.app.auth.AuthScreen
-import ai.mcpdirect.studio.app.auth.ForgotPasswordOtpVerificationScreen
-import ai.mcpdirect.studio.app.auth.ForgotPasswordScreen
-import ai.mcpdirect.studio.app.auth.LoginScreen
-import ai.mcpdirect.studio.app.auth.RegisterOtpVerificationScreen
-import ai.mcpdirect.studio.app.auth.RegisterScreen
-import ai.mcpdirect.studio.app.auth.authViewModel
+import ai.mcpdirect.studio.app.auth.*
+import ai.mcpdirect.studio.app.compose.Carousel
+import ai.mcpdirect.studio.app.compose.CarouselSlide
 import ai.mcpdirect.studio.app.dashboard.DashboardScreen
 import ai.mcpdirect.studio.app.generalViewModel
 import ai.mcpdirect.studio.app.mcpkey.MCPAccessKeyScreen
@@ -19,26 +16,28 @@ import ai.mcpdirect.studio.app.team.MCPTeamScreen
 import ai.mcpdirect.studio.app.team.MCPTeamToolMakerScreen
 import ai.mcpdirect.studio.app.team.MCPTeamToolMakerTemplateScreen
 import ai.mcpdirect.studio.app.theme.purple.PurpleTheme
+import ai.mcpdirect.studio.app.tips.QuickStartScreen
+import ai.mcpdirect.studio.app.tips.TipsScreen
 import ai.mcpdirect.studio.app.tool.MCPToolsScreen
-//import ai.mcpdirect.studio.app.tool.ToolDetailScreen
 import ai.mcpdirect.studio.app.tool.ToolPermissionScreen
 import ai.mcpdirect.studio.app.virtualmcp.VirtualMakerScreen
 import ai.mcpdirect.studio.app.virtualmcp.VirtualMakerToolConfigScreen
-import androidx.compose.foundation.background
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.safeContentPadding
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
+import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.material3.Typography
-import androidx.compose.runtime.*
+import androidx.compose.material3.VerticalDivider
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import mcpdirectstudioapp.composeapp.generated.resources.Res
+import mcpdirectstudioapp.composeapp.generated.resources.one_key_all_mcps
+import mcpdirectstudioapp.composeapp.generated.resources.openapi_mcp_tools
+import mcpdirectstudioapp.composeapp.generated.resources.share_mcp_tools
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
@@ -76,12 +75,16 @@ fun App(
                         Screen.MCPAccessKey(),
                         Screen.MCPTools,
                         Screen.MCPTeam(),
+                        Screen.Tips
 //                        Screen.ToolDevelopment -> {}
 //                        Screen.ToolsLogbook,
 //                        Screen.VirtualMCP,
                     )
                 ){
                     when (val screen = generalViewModel.currentScreen) {
+                        Screen.Tips -> {
+                            TipsScreen()
+                        }
                         Screen.Dashboard -> {
                             DashboardScreen()
                         }
@@ -129,6 +132,9 @@ fun App(
 //                        Screen.OpenAPIMCP -> {
 //                            OpenAPIToolMakerScreen()
 //                        }
+                        Screen.QuickStart ->{
+                            QuickStartScreen()
+                        }
                     }
                 }
             } else Row(
@@ -136,6 +142,43 @@ fun App(
                 horizontalArrangement = Arrangement.Center,
                 verticalAlignment = Alignment.CenterVertically
             ){
+                val slides = listOf(
+                    CarouselSlide(
+                        imageResource = Res.drawable.one_key_all_mcps, // Platform-specific
+                        description = "One API Key for your any MCP Servers in house",
+                        title = "Let MCP power your business"
+                    ),
+                    CarouselSlide(
+                        imageResource = Res.drawable.openapi_mcp_tools,
+                        description = "Zero code to connect OpenAPI as MCP Tools",
+                        title = "Let MCP power your business"
+                    ),
+                    CarouselSlide(
+                        imageResource = Res.drawable.share_mcp_tools,
+                        description = "Share MCP tools with your team",
+                        title = "Let MCP power your business"
+                    )
+                )
+                Carousel(
+                    slides,
+                    Modifier.width(800.dp),
+                )
+//                Column(
+//                    Modifier.width(800.dp),
+//                    verticalArrangement = Arrangement.Center,
+//                    horizontalAlignment = Alignment.CenterHorizontally
+//                ) {
+//                    Text(
+//                        "One API Key for your any MCP Servers in house",
+//                        style = MaterialTheme.typography.headlineMedium,
+//                    )
+//                    Image(
+//                        painterResource(Res.drawable.one_key_all_mcps),
+//                        contentDescription = null,
+//                        modifier = Modifier.width(600.dp)
+//                    )
+//                }
+                VerticalDivider()
                 when(authViewModel.currentScreen){
                     AuthScreen.Login -> {
                         LoginScreen()