+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/Screen.kt
@@ -79,4 +79,6 @@ sealed class Screen(open val title: StringResource, open val icon: DrawableResou
         val toolMaker: AIPortToolMaker?=null,
         val dialog: MyStudioScreenDialog = MyStudioScreenDialog.None
     ) : Screen(Res.string.connect_mcp, Res.drawable.plug_connect)
+    object Tips : Screen(Res.string.tips, Res.drawable.lightbulb_2)
+    object QuickStart : Screen(Res.string.tips, Res.drawable.lightbulb_2)
 }