+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/model/OpenAPIServerConfig.kt
@@ -3,11 +3,11 @@ package ai.mcpdirect.studio.app.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-class OpenAPIServerConfig {
-    var id: Long = 0
-    var name: String = ""
-    var toolPrefix:String =""
-    var status: Int = 1
+class OpenAPIServerConfig: MCPConfig() {
+//    var id: Long = 0
+//    var name: String = ""
+//    var toolPrefix:String =""
+//    var status: Int = 1
 
     var url: String? = null