+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/model/repository/StudioRepository.kt
@@ -36,6 +36,8 @@ object StudioRepository {
     ))
     val toolAgents: StateFlow<Map<Long, AIPortToolAgent>> = _toolAgents
     private val _toolMakerLastQueries = mutableMapOf<Long, TimeMark>()
+
+
     private val _mcpServers = MutableStateFlow<Map<Long, MCPServer>>(emptyMap())
     fun mcpServer(id:Long): MCPServer?{
         return _mcpServers.value[id];